+++ lib/hooks.js
@@ -5,7 +5,7 @@
  */
 
 /** @typedef {import("webpack").Compilation} WebpackCompilation */
-/** @typedef {import("../index.js")} HtmlRspackPlugin */
+/** @typedef {import("./index.js")} HtmlRspackPlugin */
 /** @typedef {import("../typings").Hooks} HtmlRspackPluginHooks */
 
 const { AsyncSeriesWaterfallHook } = require('@rspack/lite-tapable');
