+++ lib/index.js
@@ -6,22 +6,21 @@ const promisify = require('util').promisify;
 const vm = require('vm');
 const fs = require('fs');
 const path = require('path');
-const { CachedChildCompilation } = require('./lib/cached-child-compiler');
+const { CachedChildCompilation } = require('./cached-child-compiler.js');
 
 const {
   createHtmlTagObject,
   htmlTagObjectToString,
   HtmlTagArray,
-} = require('./lib/html-tags');
-const prettyError = require('./lib/errors.js');
-const chunkSorter = require('./lib/chunksorter.js');
-const getHtmlRspackPluginHooks =
-  require('./lib/hooks.js').getHtmlRspackPluginHooks;
+} = require('./html-tags.js');
+const prettyError = require('./errors.js');
+const chunkSorter = require('./chunksorter.js');
+const getHtmlRspackPluginHooks = require('./hooks.js').getHtmlRspackPluginHooks;
 
-/** @typedef {import("./typings").HtmlTagObject} HtmlTagObject */
-/** @typedef {import("./typings").Options} HtmlWebpackOptions */
-/** @typedef {import("./typings").ProcessedOptions} ProcessedHtmlWebpackOptions */
-/** @typedef {import("./typings").TemplateParameter} TemplateParameter */
+/** @typedef {import("../typings.js").HtmlTagObject} HtmlTagObject */
+/** @typedef {import("../typings.js").Options} HtmlWebpackOptions */
+/** @typedef {import("../typings.js").ProcessedOptions} ProcessedHtmlWebpackOptions */
+/** @typedef {import("../typings.js").TemplateParameter} TemplateParameter */
 /** @typedef {import("webpack").Compiler} Compiler */
 /** @typedef {ReturnType<Compiler["getInfrastructureLogger"]>} Logger */
 /** @typedef {import("webpack/lib/Compilation.js")} Compilation */
@@ -243,10 +242,8 @@ class HtmlRspackPlugin {
 
     // If the template doesn't use a loader use the lodash template loader
     if (template.indexOf('!') === -1) {
-      template =
-        require.resolve('./lib/loader.js') +
-        '!' +
-        path.resolve(context, template);
+      const loader = require.resolve('./loader.js');
+      template = loader + '!' + path.resolve(context, template);
     }
 
     // Resolve template path
